fix(2D images): add hover info to 2D cells and layouts.
Use a tooltip to show cell radii and materials, or which cell is
in a grid position.

Requires a patch to tooltip library to avoid static empty tooltip
when hovering outside the cell image outer radius. Not sure how
to deliver this yet.

Make new materials be based off the current material.